Move schedule types to separate file
